Move timedcache into own module
